[MODIFY]ファイルをmain.goから分割
